Revert "refactored websocket"
This reverts commit becbda7773725e3ae2a29e874fdca398623518d6.
