--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Data.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 56781367b49846bd9ddaf869dc735ff1
-timeCreated: 1649418698
