--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Data/WebSocketCloseCode.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 1aca653236714d7abbe844d8db71af2a
-timeCreated: 1649417883
