--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Data/WebSocketState.cs
@@ -1,10 +1,0 @@
-namespace AnkrSDK.WalletConnectSharp.Unity.Network.Client.Data
-{
-	public enum WebSocketState
-	{
-		Connecting,
-		Open,
-		Closing,
-		Closed
-	}
-}
