--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Data/WebSocketState.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 86cb7ed890624a24b9e7b87b8f4d7ca3
-timeCreated: 1649417889
