--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/EvenHandlers.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: ca4a154183f34b298e243a94a51aecfa
-timeCreated: 1649421132
