--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/EvenHandlers/WebSocketErrorEventHandler.cs
@@ -1,4 +1,0 @@
-namespace AnkrSDK.WalletConnectSharp.Unity.Network.Client.EvenHandlers
-{
-	public delegate void WebSocketErrorEventHandler(string errorMsg);
-}
