--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/EvenHandlers/WebSocketErrorEventHandler.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: a58d9b337d7448d082f9e781b98659f5
-timeCreated: 1649418746
