--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/EvenHandlers/WebSocketMessageEventHandler.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 5bbb79a710e44a78ac450c26579d82a1
-timeCreated: 1649418743
