--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/EvenHandlers/WebSocketOpenEventHandler.cs
@@ -1,4 +1,0 @@
-namespace AnkrSDK.WalletConnectSharp.Unity.Network.Client.EvenHandlers
-{
-	public delegate void WebSocketOpenEventHandler();
-}
