--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/EvenHandlers/WebSocketOpenEventHandler.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: d7c4f3eeca894ded9e9172f799190b03
-timeCreated: 1649418739
