--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Exceptions.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 6c3bad3455c5400dabba68e80895fc03
-timeCreated: 1649418665
