--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Exceptions/WebSocketException.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 39b68d95948f44ed80d8bb76ecedf006
-timeCreated: 1649417963
