--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Exceptions/WebSocketInvalidArgumentException.cs
@@ -1,19 +1,0 @@
-using System;
-
-namespace AnkrSDK.WalletConnectSharp.Unity.Network.Client.Exceptions
-{
-	public class WebSocketInvalidArgumentException : WebSocketException
-	{
-		public WebSocketInvalidArgumentException()
-		{
-		}
-
-		public WebSocketInvalidArgumentException(string message) : base(message)
-		{
-		}
-
-		public WebSocketInvalidArgumentException(string message, Exception inner) : base(message, inner)
-		{
-		}
-	}
-}
