--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Exceptions/WebSocketInvalidArgumentException.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 57b379773a94410aad2adaf20583662d
-timeCreated: 1649417972
