--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Exceptions/WebSocketInvalidStateException.cs
@@ -1,19 +1,0 @@
-using System;
-
-namespace AnkrSDK.WalletConnectSharp.Unity.Network.Client.Exceptions
-{
-	public class WebSocketInvalidStateException : WebSocketException
-	{
-		public WebSocketInvalidStateException()
-		{
-		}
-
-		public WebSocketInvalidStateException(string message) : base(message)
-		{
-		}
-
-		public WebSocketInvalidStateException(string message, Exception inner) : base(message, inner)
-		{
-		}
-	}
-}
