--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Exceptions/WebSocketInvalidStateException.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: dffdeef90417415988fd7ff683375d88
-timeCreated: 1649417977
