--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Exceptions/WebSocketUnexpectedException.cs
@@ -1,19 +1,0 @@
-using System;
-
-namespace AnkrSDK.WalletConnectSharp.Unity.Network.Client.Exceptions
-{
-	public class WebSocketUnexpectedException : WebSocketException
-	{
-		public WebSocketUnexpectedException()
-		{
-		}
-
-		public WebSocketUnexpectedException(string message) : base(message)
-		{
-		}
-
-		public WebSocketUnexpectedException(string message, Exception inner) : base(message, inner)
-		{
-		}
-	}
-}
