--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Exceptions/WebSocketUnexpectedException.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: ac7cbc3781164e1f9969623fd338812c
-timeCreated: 1649417967
