--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Helpers.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: c422dad14f3f48e2b49d974163748dbc
-timeCreated: 1649418650
