--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Helpers/WaitForBackgroundThread.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 34ea30ea835e447fb7257fc536e515b0
-timeCreated: 1649417982
