--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Helpers/WebSocketFactory.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: c72c10b45ea248df950c7033c993079a
-timeCreated: 1649440133
