--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Helpers/WebSocketHelpers.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 667616ca458444cab0df85e59aeeaf83
-timeCreated: 1649417957
