--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Implementation.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 71c14269cde34193bd9f51faf14e8f2f
-timeCreated: 1649418022
