--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Implementation/WebGLWebSocket.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 01ba860b2379410e9c98afb173066cd7
-timeCreated: 1649418233
