--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Implementation/WebSocket.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 80fdf43e3130473fa13ba9525db878cb
-timeCreated: 1649418030
