--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Infrastructure.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 8da1c884680c4569ae1922f70a937f27
-timeCreated: 1649418007
