--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Infrastructure/IWebSocket.cs
@@ -1,21 +1,0 @@
-using System.Threading.Tasks;
-using AnkrSDK.WalletConnectSharp.Unity.Network.Client.Data;
-using AnkrSDK.WalletConnectSharp.Unity.Network.Client.EvenHandlers;
-
-namespace AnkrSDK.WalletConnectSharp.Unity.Network.Client.Infrastructure
-{
-	public interface IWebSocket
-	{
-		event WebSocketOpenEventHandler OnOpen;
-		event WebSocketMessageEventHandler OnMessage;
-		event WebSocketErrorEventHandler OnError;
-		event WebSocketCloseEventHandler OnClose;
-
-		WebSocketState State { get; }
-		void CancelConnection();
-		void DispatchMessageQueue();
-		Task Close();
-		Task Connect();
-		Task SendText(string message);
-	}
-}
