--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/Infrastructure/IWebSocket.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 014c262256e943a8b5a61b839897cfd3
-timeCreated: 1649417948
