+++ Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/NativeWebSocket.cs.meta
