--- Assets/AnkrSDK/Plugins/WalletConnectSharp.Unity/Network/Client/WebSocketNativeBridge.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: f1a982f2e3514fcdb66d183d0b105bba
-timeCreated: 1649418311
