+++ Assets/AnkrSDK/Scripts/Core/Implementation/ContractEventSubscriber.cs
@@ -3,9 +3,8 @@ using System.Collections.Generic;
 using AnkrSDK.Core.Data;
 using AnkrSDK.Core.Infrastructure;
 using AnkrSDK.Core.Utils;
-using AnkrSDK.WalletConnectSharp.Unity.Network.Client.Data;
-using AnkrSDK.WalletConnectSharp.Unity.Network.Client.Implementation;
 using Cysharp.Threading.Tasks;
+using NativeWebSocket;
 using Nethereum.ABI.FunctionEncoding.Attributes;
 using Nethereum.JsonRpc.Client;
 using Nethereum.JsonRpc.Client.RpcMessages;
